refactor: countOfArticles 는 Long 타입으로 countOfReplies는 Integer타입으로 변경

+++ src/main/java/kr/codesqaud/cafe/article/ArticleService.java
@@ -67,7 +67,7 @@ public class ArticleService {
 		}
 	}
 
-	public int getCountOfArticles() {
+	public Long getCountOfArticles() {
 		return articleRepository.getCountOfArticles();
 	}
 }
